% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr11_remove_genes}
\alias{dr11_remove_genes}
\title{Genes to Remove from ScRNA-seq Analysis:  Zebrafish}
\format{
A character vector
}
\source{
Ensembl
}
\usage{
dr11_remove_genes
}
\description{
A precompiled list of mitochondrial and ribosomal gene names to remove.
}
\details{
See data-raw/remove_genes
}
\keyword{datasets}
