% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr11_tss}
\alias{dr11_tss}
\title{Zebrafish Gene TSS}
\format{
A GRanges object:
}
\usage{
dr11_tss
}
\description{
From TxDb.Drerio.UCSC.danRer11.refGene
}
\details{
see data-raw/tss.R
}
\keyword{datasets}
