% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drerio}
\alias{Drerio}
\title{Drerio}
\format{
A BSgenome object:
}
\source{
UCSC danRer11 genome build
}
\usage{
Drerio
}
\description{
This is a copy of the Drerio object from BSgenome.Drerio.UCSC.danRer11.  It is here for more convenient use with blaseRtools functions and to reduce dependencies.
}
\details{
see data-raw/grcz11_objects.R
}
\keyword{datasets}
