% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_granges}
\alias{hg38_granges}
\title{Human Gene Model}
\format{
A GRanges object:
}
\source{
http://ftp.ensembl.org/pub/release-104/gff3/homo_sapiens/
}
\usage{
hg38_granges
}
\description{
Full Genome Model for humans.  Based on GRCh38 genome build.
}
\details{
see data-raw/hg38_objects.R
}
\keyword{datasets}
