% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_granges_reduced}
\alias{hg38_granges_reduced}
\title{Reduced Human Gene Model}
\format{
A GRanges object:
}
\source{
http://ftp.ensembl.org/pub/release-104/gff3/homo_sapiens/
}
\usage{
hg38_granges_reduced
}
\description{
Similar to the Human Gene Model "hg38_granges" but with different metadata columns and only with 5', CDS, and 3' UTR ranges.  Also has principle isoform data.
}
\details{
see data-raw/trace_objects.R
}
\keyword{datasets}
