% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_promoters}
\alias{hg38_promoters}
\title{Human Gene Promoters}
\format{
A GRanges object:
}
\source{
http://ftp.ensembl.org/pub/release-104/gff3/homo_sapiens/
}
\usage{
hg38_promoters
}
\description{
Defined as TSS +/- 200 bp.  TSS identified as the start of each mRNA in hg38_granges according to strand. (start of the GRange for top strand and end of the GRange for bottom strand.)
}
\details{
see data-raw/tss.R
}
\keyword{datasets}
