% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_tss}
\alias{hg38_tss}
\title{Human Gene TSS}
\format{
A GRanges object:
}
\usage{
hg38_tss
}
\description{
From TxDb.Hsapiens.UCSC.hg38.knownGene
}
\details{
see data-raw/tss.R
}
\keyword{datasets}
