% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hsapiens}
\alias{Hsapiens}
\title{Hsapiens}
\format{
A BSgenome object:
}
\source{
UCSC hg38 genome build
}
\usage{
Hsapiens
}
\description{
This is a copy of the Hsapiens object from BSgenome.Hsapiens.UCSC.hg38.  It is here for more convenient use with blaseRtools functions and to reduce dependencies.
}
\details{
see data-raw/hg38_objects.R
}
\keyword{datasets}
