% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meme}
\alias{meme}
\title{Jaspar 2020 TF Motifs}
\format{
A character vector.
}
\source{
JASPAR2020_CORE_vertebrates_non-redundant_pfms_jaspar.txt
}
\usage{
meme
}
\description{
JASPAR 2020 TF Motif database in character string format.  Mostly for internal use.
}
\details{
see data-raw/meme.R
}
\keyword{datasets}
