% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_genesets}
\alias{msigdb_genesets}
\title{Broad Institute MSIGDB Gene Sets}
\format{
A list.
}
\source{
https://www.gsea-msigdb.org/gsea/msigdb/ accessed January 12 2022.
}
\usage{
msigdb_genesets
}
\description{
List of MSIGDB gene sets.  Most gene sets are commonly known by "STANDARD_NAME".  This is a unique value and is provided as the name for each of the elements in this list. If you want to select a single gene set by name you can use the usual R syntax for selecting named elements:  msigdb_genesets$<gene set name>.  If you want to select by other metadata features you should use bb_extract_msig.
}
\details{
Use with bb_extract_msig
}
\keyword{datasets}
