% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vig_qc_res}
\alias{vig_qc_res}
\title{Vignette QC Results}
\format{
a list
}
\source{
vignette_cds
}
\usage{
vig_qc_res
}
\description{
Standard bb_qc function output
}
\details{
See blaseRtools scRNAseq Vignette
}
\keyword{datasets}
