% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_cds}
\alias{vignette_cds}
\title{ScRNA-seq Vignette CDS}
\format{
A cellDataSet object
}
\source{
https://www.10xgenomics.com/resources/datasets/500-human-pbm-cs-3-lt-v-3-1-chromium-controller-3-1-low-6-1-0 and https://www.10xgenomics.com/resources/datasets/500-human-pbm-cs-3-lt-v-3-1-chromium-x-3-1-low-6-1-0
}
\usage{
vignette_cds
}
\description{
A small example CDS for the blaseRtools vignette
}
\details{
See blaseRtools scRNAseq Vignette
}
\keyword{datasets}
