% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_pseudobulk_res}
\alias{vignette_pseudobulk_res}
\title{BlaseRtools ScRNA-seq Vignette CDS Pseudobulk Results}
\format{
A list of pseudobulk results
}
\usage{
vignette_pseudobulk_res
}
\description{
Pseudobulk header, results and qc plots.
}
\details{
See blaseRtools scRNAseq Vignette
}
\keyword{datasets}
