% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_top_markers}
\alias{vignette_top_markers}
\title{BlaseRtools ScRNA-seq Vignette CDS Top Markers}
\format{
A data frame with 1100 rows and 11 variables:
\describe{
\item{\code{gene_id}}{character Ensembl gene identifier}
\item{\code{gene_short_name}}{character Gene Name}
\item{\code{cluster_method}}{character Clustering Method}
\item{\code{cell_group}}{character Cell Cluster or Partition}
\item{\code{marker_score}}{double Composite score of fraction expressing and level of expression.}
\item{\code{mean_expression}}{double Mean marker expression}
\item{\code{fraction_expressing}}{double Fraction expressing}
\item{\code{specificity}}{double Marker specificity}
\item{\code{pseudo_R2}}{double Regression pseudo R squared}
\item{\code{marker_test_p_value}}{double Wald test P value}
\item{\code{marker_test_q_value}}{double q value}
}
}
\usage{
vignette_top_markers
}
\description{
Top Markers for the vignette CDS
}
\details{
See blaseRtools scRNAseq Vignette
}
\keyword{datasets}
