% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordhash}
\alias{wordhash}
\title{Wordlist Hashtable}
\format{
A data frame with 226857 rows and 2 variables:
\describe{
\item{\code{index}}{integer }
\item{\code{word}}{character }
}
}
\source{
https://www.gutenberg.org/ebooks/3201
}
\usage{
wordhash
}
\description{
Wordlist from the Grady Parts of Speech database.  In order to reuse, 1.) Calculate the md5sum of the object you want to hash, 2.) Take the first 5 or more hex digits and fill them in this expression:  as.integer(as.hexmode("<hex digits>") \%\% 226857) + 1.  This will return an integer between 1 and 226857 which is the length of the wordlist hashtable.  Then select the word according to the integer/index value using something like:  wordlist \%>\% filter(index == integer) \%>\% pull(word).
}
\details{
see data-raw/wordhash.R for construction
}
\keyword{datasets}
