% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zfin_granges}
\alias{zfin_granges}
\title{Zebrafish Gene Model}
\format{
A GRanges object:
}
\source{
https://zfin.org/downloads/zfin_genes.gff3 release April 2018
}
\usage{
zfin_granges
}
\description{
This is a granges object containing the full genome model from the zfin database.  Based on GRCz11 genome build.
}
\details{
see data-raw/grcz11_objects.R
}
\keyword{datasets}
