% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zfin_granges_reduced}
\alias{zfin_granges_reduced}
\title{Reduced Zebrafish Gene Model}
\format{
A GRanges object:
}
\source{
https://zfin.org/downloads/zfin_genes.gff3 release April 2018
}
\usage{
zfin_granges_reduced
}
\description{
Similar to the Zebrafish Gene Model but with different metadata columns and only with 5', CDS and 3' UTR ranges.  Also has principle isoform data.
}
\details{
see data-raw/trace_objects.R
}
\keyword{datasets}
