% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zfin_promoters}
\alias{zfin_promoters}
\title{Zebrafish Gene Promoters}
\format{
A GRanges object:
}
\source{
https://zfin.org/downloads/zfin_genes.gff3 release April 2018
}
\usage{
zfin_promoters
}
\description{
Defined as TSS +/- 200 bp.  TSS identified as the start of each mRNA in zfin_granges according to strand. (start of the GRange for top strand and end of the GRange for bottom strand.)
}
\details{
see data-raw/tss.R
}
\keyword{datasets}
